Don't force subclass to include all parent's tests
When determing what tests to run for a class, loader.py previously
would walk the class's base classes and include any test that had
the same name as a test being run for a base class.

There is no reason to do this and it leads to the wrong behavior
in cases where the parent should run the test but not the child,
for example when the child has a test with the same name but
different attributes and the attrib plugin is being used.

Fixes #464.