xunit: capture stdout out into the system-out tag
Change the score of Xunit to place it before the Capture plugin.  We
don't want the Capture plugin to change the streams before the tests are
run.  At a score of 2000, that's exactly what happens, so let's make it
499 instead (Capture's score - 1).

Have Xunit replace the stream with a Tee object of just swapping the
stream.  This allows the Capture plugin to see the data from stdout,
even though we've replaced the stream, allowing it to function exactly
as you would expect it to.

Finally, we maintain a stack of streams, instead of just a single
stream.  I don't believe this is entirely necessary though.