multiprocess: make test a little less racy
The other process can start up not long after the module setup and
teardown is called.  Add another ellipsis in to cover this case.
Note: it's feasible that something could sneak in between the setup and
teardown as well, but the failures with Travis haven't shown that to be
occurring.