Fix --cover-min-percentage with single packages.
It turns out that coverage doesn't generate a TOTAL line when only one
package is being covered.  As a result, we'd fail to parse the output
and assert accordingly.  Fix this by first trying a multi-package
regexp, and fall back to a single package version if the multi-package
one fails.  Also added several tests to make sure this works with and
without --cover-branches.