Correctly detect the coverage module under Python 3.3.
Python 3.3 has namespace packages (PEP 420) and the coverage support
area is being picked up as one.  As a result, when trying to run tests,
the current infrastructure believes that coverage is installed when it's
not, leading to several test failures.  Let's do an additional check to
make sure we have the real coverage package installed.