A slightly better fix for #627.
The first test should run fine, even if the number of processors is one.
That's because we explicitly tell multiprocess to use two process.
However, the second test asks multiprocess for the CPU count, and that's
where we run into a problem.  As a result, we'll only check the cpu
count in TestConcurrentSharedWithAutomaticProcessesCount.