Fix include directive in autoplugin_directive.
It appears that the include directive does not like absolute paths, and
ends up stripping the leading '/'.  This means that it fails to find the
python module, and warns in the Sphinx build.

Instead of using an absolute path, let's formulate a relative path using
docutils.utils.relative_path().  Since the path needs to be relative to
the document, let's pull that from state_machine.document['source'].