Fix #692: UnicodeDecodeError in xunit when capturing stdout and stderr
It turns out that `Xunit` was very mix-brained about how to formulate
the blocks ultimately used for generating the report.  In some parts,
the expectation was to put unicode strings into the blocks, and in
others it would insert encoded strings along with unicode strings into
the blocks.  It's the latter that was ultimately causing the issue.

In particular, `_quoteattr` would sometimes receive a unicode string,
and then it would get translated into UTF-8.  Then, it would get
inserted into a unicode string, which would cause Python to try and turn
the UTF-8 encoded string back into a unicode one, but it would default
to the ascii codec and fail.

This makes us use unicode everywhere, sometimes forcing the data into
unicode with replacement.