setup.py: Don't create man files
GitHub Issue #731 explains an error that does not allow nose to be installed
globally to a device. Although it was determined that this error might be due
to setuptools (or derivates), it was not fixed yet and the fundamental problem
remains that nose is not installable in a system global context without
virtualenvs.

Therefore this commit removes setup.py's ability to create manual files under
the assumption that most people using virtualenv will not use the man file but
instead refer to the web documentation anyway. Thus most people should not even
realise the missing man files, while in exchange people can now install nose
without virtualenvs.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>