Fix #728: attrib plugin rejects any staticmethod
The issue is that static methods get wrapped in an UnboundMethod object,
and the attribute plugin is unable to grab the associated class, since
UnboundMethod does not have the im_class attribute.  Since the goal of
UnboundMethod is to mimic an actual method, let's go ahead and implement
the im_class attribute.