Fix #783: try_run is broken with Python 3.4
It turns out that we expected inspect.getargspec() to only work on
functions (versus classes that are callable).  Python 3.4 has changed
this behavior and now happily returns the arg spec.  Let's try to detect
this situation by checking the __call__() method to if it's actually the
method of a class.  If so, pop the 'self' argument off of args--similar
to what's done in the exceptional case below.