Fix #815: "ValueError: unsupported pickle protocol" with --with-id...
This would occur by running `nosetests --with-id` with a newer version
of Python, and then run again with an older version of Python who
doesn't recognize the pickle format.  In this case, let's simply ignore
the file, and move on allowing nosetests to overwrite it after gathering
the data again.