Fix #875: nose doesn't collect tests when subpackage is given as arg
When a subpackage is specified on the command line (e.g., `nosetests
foo.bar`), it should pick up all the tests below the subpackage.

The root of the problem came from how we resolved a filename for a
package.  We should really point to the package directory, instead of
providing the path to the `__init__.py` file.  Otherwise, we only end up
attempting to select modules from the `__init__.py` instead of the whole
tree underneath.  We don't want to check for a path ending in
`__init__.py` in `loadTestsFromModule()` because then we select the
whole subtree of tests when `foo/bar/__init__.py` is specified when only
tests in the __init__.py should be selected.