unit_tests/test_cases: add tests for {Function,Method}TestCase.__repr__
At present, if a mutable argument is changed during the execution of test
case, then any printing of the test case object which occurs after the
test has run will result in the argument being printed with a __repr__ of
it's *new* state (assuming that its __repr__ is state-dependant). This is
particularly confusing when debugging failing test cases because the
summary lines and the failure-detail lines will show *different* __repr__s
of the args, e.g.

  nose_weirdness.test_nose_is_weird(Foo(10), 0) ... FAIL

  ======================================================================
  FAIL: nose_weirdness.test_nose_is_weird(Foo(0), 0)
  ----------------------------------------------------------------------
  ...
  Ran 1 test in 0.001s

  FAILED (failures=1)

This issue will be solved in the next commit.