Fix #931: duplicate __init__ modules can be picked up
The short form here is that when setting up ignores, you may allow
__init__.py to be included unexpectedly.  Since we generally use
loadTestsFromName(), the package should be imported first which will
handle the __init__.py case.  Therefore, let's always ignore it in the
loadTestsFromDir() method.  Just to make sure that we ignore the correct
files, we need to run the entry name through src() to make sure we
handle the situation correctly on other platforms (like Jython).