case.{Function,Method}TestCase: deobfuscate references to self.arg_repr
Previously, both `FunctionTestCase` and `MethodTestCase` created local
`arg` variables in their `__str__` methods. These variables were simply
references to `self.arg_repr`, so this commit deobfuscates this reference
by modifying the `_descriptors` method for each so that *only the test
name* is returned and not the `arg_repr` also (renaming `_descriptors` to
`_descriptor` in the process), and then referring to `self.arg_repr`
directly.

This makes the `__str__` easier to understand because it is immediately
obvious that we're only dealing with the string-representation of the args
and *not* the actual args themselves.